refactor: DetailPage 분리
